add NetIOCountersRateHandler
