+++ src/handlers_psutil.py
@@ -2,14 +2,13 @@
 # Licensed under the MIT License.  See LICENSE file in the project root for full license information.
 
 import psutil
-import time
 from typing import (
     NamedTuple,
 )
 
-from .handlers_base import BaseHandler, MethodCommandHandler, NameOrTotalTupleCommandHandler, Payload
+from .handlers_base import MethodCommandHandler, NameOrTotalTupleCommandHandler, Payload, TaskParam
+from .handlers_derived import RateHandler
 from .utils import string_from_dict_optionally
-from .handlers_base import TaskParam
 
 class DiskIOCountersCommandHandler(MethodCommandHandler):
     '''
@@ -85,99 +84,15 @@ class DiskIOCountersCommandHandler(MethodCommandHandler):
             raise Exception(f"{self.name}: Disk '{disk}' is not valid. Available disks: {avail_disks}")
 
 
-class DiskIOCountersRateHandler(BaseHandler):
+class DiskIOCountersRateHandler(RateHandler):
     '''
     DiskIOCountersRateHandler computes the rate of change of the disk I/O counters.
     This is often more useful than the monotonically-increasing raw disk I/O counters.
     '''
 
-    MINIMAL_DELTA_TIME_SECONDS = 0.1
-
     def __init__(self) -> None:
-        super().__init__('disk_io_counters_rate')
-        self.monotonic_counter_handler = DiskIOCountersCommandHandler()
-        self.last_values = {}
-        self.last_timestamp = {}
+        super().__init__('disk_io_counters_rate', DiskIOCountersCommandHandler())
         return
-
-    @staticmethod
-    def compute_rate_from_dicts(new_values: dict, last_values: dict, delta_time_seconds: float) -> dict:
-        # compute the rate of change of the counters
-        result = {}
-        for k in new_values.keys():
-            if k in last_values:
-                # IMPORTANT: no checks are done on the result being negative... ideally this should never happen
-                #            unless psutil has some internal counter reset and its monotonically increasing counters
-                #            happen to decrease. This is not expected to happen in normal operation.
-                result[k] = int((new_values[k] - last_values[k]) / delta_time_seconds)
-            else:
-                result[k] = new_values[k]
-
-        return result
-
-    @staticmethod
-    def compute_rate_from_tuples(new_values: tuple, last_values: tuple, delta_time_seconds: float) -> dict:
-        return DiskIOCountersRateHandler.compute_rate_from_dicts(new_values._asdict(), last_values._asdict(), delta_time_seconds)
-
-    def handle(self, params: list[str], caller_task_id: str) -> Payload:
-        if caller_task_id not in self.last_values:
-            # this is the first sample being retrieved... just save the current values
-            # and we'll be able to compute the rate/delta of the next call
-            new_values = self.monotonic_counter_handler.handle(params, caller_task_id)
-            new_timestamp = time.time()
-
-            # we return zero(s) on this first sample to avoid pushing a HUGE absolute value
-            # which might decrease nearly to zero on the next sample
-            if isinstance(new_values, dict):
-                result = {k: 0 for k in new_values.keys()}
-            elif isinstance(new_values, tuple):
-                result = (0,) * len(new_values)
-            elif isinstance(new_values, int):
-                result = 0
-            else:
-                raise Exception(f"{self.name}: Unexpected result type: {type(new_values)}")
-
-            #logging.debug(f"{self.name}: producing first sample as zeroes for caller task {caller_task_id}: {result}")
-
-            # fall to the end of the function where we update internal state
-        else:
-            # retrieve previous values / timestamp -- we know they are in the internal state
-            old_values = self.last_values[caller_task_id]
-            old_timestamp = self.last_timestamp[caller_task_id]
-
-            # get the new sensor readings:
-            new_values = self.monotonic_counter_handler.handle(params, caller_task_id)
-            new_timestamp = time.time()
-
-            delta_time_seconds = new_timestamp - old_timestamp
-            if delta_time_seconds <= DiskIOCountersRateHandler.MINIMAL_DELTA_TIME_SECONDS:
-                # delta is too small... return the last value and skip any internal update
-                return old_values
-
-            #logging.debug(f"{self.name}: computing rate with delta_time_seconds={delta_time_seconds} for caller task {caller_task_id}")
-
-            if isinstance(new_values, dict):
-                assert isinstance(old_values, dict)
-                result = DiskIOCountersRateHandler.compute_rate_from_dicts(new_values, old_values, delta_time_seconds)
-            elif isinstance(new_values, tuple):
-                assert isinstance(old_values, tuple)
-                result = DiskIOCountersRateHandler.compute_rate_from_tuples(new_values, old_values, delta_time_seconds)
-            elif isinstance(new_values, int):
-                assert isinstance(old_values, int)
-                # IMPORTANT: no checks are done on the result being negative... ideally this should never happen
-                #            unless psutil has some internal counter reset and its monotonically increasing counters
-                #            happen to decrease. This is not expected to happen in normal operation.
-                result = int((new_values - old_values) / delta_time_seconds)
-            else:
-                raise Exception(f"{self.name}: Unexpected result type: {type(new_values)}")
-
-        # update internal state (by caller task)
-        self.last_values[caller_task_id] = new_values
-        self.last_timestamp[caller_task_id] = new_timestamp
-        return result
-
-    def get_value(self) -> Payload:
-        raise Exception("This method should not be called")
 
 class DiskUsageCommandHandler(MethodCommandHandler):
     '''
@@ -225,6 +140,17 @@ class NetIOCountersCommandHandler(NameOrTotalTupleCommandHandler):
 
     def get_value(self, total:bool) -> Payload:
         return psutil.net_io_counters(pernic=not total)
+
+
+class NetIOCountersRateHandler(RateHandler):
+    '''
+    NetIOCountersRateHandler computes the rate of change of the network I/O counters.
+    This is often more useful than the monotonically-increasing raw network I/O counters.
+    '''
+
+    def __init__(self) -> None:
+        super().__init__('net_io_counters_rate', NetIOCountersCommandHandler())
+        return
 
 
 class SensorsTemperaturesCommandHandler(MethodCommandHandler):
