+++ src/task.py
@@ -15,7 +15,7 @@ from .formatter import Formatter
 
 from .handlers_base import Payload, TupleCommandHandler, ValueCommandHandler, IndexCommandHandler, IndexOrTotalCommandHandler, IndexTupleCommandHandler, IndexOrTotalTupleCommandHandler
 from .handlers_psutil_processes import ProcessesCommandHandler
-from .handlers_psutil import DiskIOCountersCommandHandler, DiskIOCountersRateHandler, DiskUsageCommandHandler, NetIOCountersCommandHandler, SensorsFansCommandHandler, SensorsTemperaturesCommandHandler
+from .handlers_psutil import DiskIOCountersCommandHandler, DiskIOCountersRateHandler, DiskUsageCommandHandler, NetIOCountersCommandHandler, NetIOCountersRateHandler, SensorsFansCommandHandler, SensorsTemperaturesCommandHandler
 from .handlers_pysmart import SmartCommandHandler
 
 class Task:
@@ -76,6 +76,7 @@ class Task:
         # NETWORK
 
         'net_io_counters': NetIOCountersCommandHandler(),
+        'net_io_counters_rate': NetIOCountersRateHandler(),
 
         # PROCESSES
